FIX: Service, Repository 리팩토링 - 메서드 명 변경 (Cert -> One, List, Page 로 반환 값 메서드에 구분해서 표기) - 중복 코드 제거 - 조회용 서비스 Query, 변경용 서비스 Command로 구분

+++ src/main/java/com/delgo/reward/dto/map/OtherMapDTO.java
@@ -19,10 +19,10 @@ public class OtherMapDTO {
     private int viewCount;
 
     List<CertResponse> certs;
-    int totalCount;
+    long totalCount;
 
 
-    public OtherMapDTO(User user, List<CertResponse> certs, int totalCount){
+    public OtherMapDTO(User user, List<CertResponse> certs, long totalCount){
         userId = user.getUserId();
         nickname = user.getName();
         viewCount = user.getViewCount();
